Save: 작업환경 옮김으로 현재 작업 상태 저장
