+++ src/app/page.tsx
@@ -1,6 +1,7 @@
 import RootLayout from "./layout";
 import Introduction from "./component/intro/Introduction";
 import CommonBenefits from "./component/membershipBenefits/CommonBenefits";
+import MembershipBenefits from "./component/membershipBenefits/MembershipBenefits";
 import MembershipPrice from "./component/membershipPrice/MembershipPrice";
 import { useEffect } from "react";
 // import * as styles from "./MainPage.css";
@@ -14,6 +15,7 @@ const MainPage = () => {
     <RootLayout>
       <Introduction />
       <CommonBenefits />
+      <MembershipBenefits />
       <MembershipPrice />
     </RootLayout>
   );
